cmd/install-wsl: split tests to match package structure
Split unit tests from monolithic main_test.go into domain-specific test files:

- version_test.go: TestParseVersion, TestCompareVersions
- wsl_test.go: TestWriteWslConfig
- download_test.go: TestDownloadFile, TestDownloadFileAtomicRename
- helpers_test.go: Test helper functions (contains, stringContains)

Test organization now mirrors package organization:
- version.go/version_test.go
- wsl.go/wsl_test.go
- download.go/download_test.go
- main.go (no tests yet)
- github.go (no tests yet)

All tests pass.
